Move `themes` out of `tuilib` * It is not handled by the TUI framework
